pyproject.toml
[tool.poetry]
name = "pokebotdiscord"
version = "0.1.0"
description = ""
authors = ["Fernandez Rodrigo <rdr.rkh+github2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
discord-py = "^2.4.0"

[tool.poetry.scripts]
my-script = "main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
